% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelecount.R
\name{alleleCount}
\alias{alleleCount}
\title{Run alleleCount}
\usage{
alleleCount(locifile, bam, outfile, min_baq = 20, min_maq = 35)
}
\arguments{
\item{locifile}{A file with at least chromsome and position columns of the locations to be counted}

\item{bam}{A bam file}

\item{outfile}{Where to store the output}

\item{min_baq}{The minimum base quality required for a read to be counted}

\item{min_maq}{The minimum mapping quality required for a read to be counted}
}
\description{
Count the alleles for specified locations in the loci file. Expects alleleCount binary in $PATH
}
\author{
sd11
}
