% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumber.R
\name{ascatNgsToBattenberg}
\alias{ascatNgsToBattenberg}
\title{Transform ASCAT NGS output into Battenberg}
\usage{
ascatNgsToBattenberg(
  outfile.prefix,
  copynumber.caveman.file,
  samplestatistics.file
)
}
\arguments{
\item{outfile.prefix}{String prefix for the output files. Internally _subclones.txt and _rho_and_psi.txt will be added.}

\item{copynumber.caveman.file}{String pointing to an ASCAT NGS copynumber.caveman.csv file}

\item{samplestatistics.file}{String point to an ASCAT NGS samplestatistics.csv file}
}
\description{
This function takes the ascat copynumber.caveman.csv and samplestatistics files to create
a minimum Battenberg subclones and rho_and_psi file for use in pre-processing.
These files only contain the essentials required by this package.
}
\author{
sd11
}
