% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ascatToBattenberg}
\alias{ascatToBattenberg}
\title{Transform ASCAT output into Battenberg}
\usage{
ascatToBattenberg(outfile.prefix, segments.file, acf.file, ploidy.file)
}
\arguments{
\item{outfile.prefix}{String prefix for the output files. Internally _subclones.txt and _rho_and_psi.txt will be added.}

\item{segments.file}{String pointing to a segments.txt ASCAT output file}

\item{acf.file}{String pointing to a acf.txt ASCAT output file}

\item{ploidy.file}{String pointing to a ploidy.txt ASCAT output file}
}
\description{
This function takes the ascat segments, acf and ploidy files to create
a minimum Battenberg subclones and rho_and_psi file for use in pre-processing.
These files only contain the essentials required by this package.
}
\author{
sd11
}

