% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calc_power}
\alias{calc_power}
\title{Calculate power to call subclones.}
\usage{
calc_power(purity, ploidy, coverage)
}
\arguments{
\item{purity}{The sample purity}

\item{ploidy}{The tumour ploidy}

\item{coverage}{The tumour sample coverage}
}
\value{
The average reads per chromosome copy, a.k.a. power
}
\description{
This function calculates the 
average number of reads per clonal chromosome copy.
}
\author{
sd11
}
