% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/copynumber.R
\name{collate_bb_subclones}
\alias{collate_bb_subclones}
\title{This function takes a Battenberg subclones.txt file, pulls out the A copy number profile,
then it calculates the total CN and a weighted ploidy before it subsets the input data
by columns: "chr", "startpos", "endpos", "nMaj1_A", "nMin1_A", "frac1_A",
"nMaj2_A", "nMin2_A", "frac2_A", "SDfrac_A".}
\usage{
collate_bb_subclones(samplename, bb_subclones_file, is_male, outfile)
}
\arguments{
\item{samplename,}{A String containing the samplename}

\item{bb_subclones_file,}{A String containing the full path to a Battenberg subclones.txt file}

\item{is_male,}{Boolean that should be TRUE if this sample is male, FALSE otherwise}

\item{outfile,}{A String with the full path to where the output should be written}
}
\description{
Then logic is applied for classifying copy number aberrations into
a series of classes. There are two main groups: Clonal (starting with c) and
Subclonal (starting with s).
}
\details{
A CNA can be:
* HD : Homozygous deletion
* LOH : Loss of heterozygosity
* Amp : Amplification
* Gain : Gain
* NoCNV : Normal copy number without aberrations
* Loss : Loss

The copy number segments, classification, weighted ploidy and samplename are saved
to the specified output file
}
\author{
tjm
}

