% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{concat_files}
\alias{concat_files}
\title{Concatenate split files}
\usage{
concat_files(fofn, inputdir, outfile, haveHeader)
}
\arguments{
\item{fofn}{A file of file names to be concatenated}

\item{inputdir}{Full path to where the input files are stored}

\item{outfile}{Full path to where the output should be written}

\item{haveHeader}{Boolean that specifies whether the input files have a header}
}
\description{
Convenience function to concatenate a series of files specified in a file of file names.
This function assumes all files have the same layout.
}
\author{
sd11
}

