% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qualitycontrol.R
\name{createPng}
\alias{createPng}
\title{Create a PNG file}
\usage{
createPng(p, filename, width, height)
}
\arguments{
\item{p}{The figure}

\item{filename}{Where to save the image}

\item{width}{Width}

\item{height}{Height}
}
\description{
Create a PNG file
}

