% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dpclust3p-package.R
\docType{package}
\name{dpclust3p}
\alias{dpclust3p}
\alias{dpclust3p-package}
\title{Dirichlet Process Clustering PreProcessing Package}
\description{
Dirichlet Process Clustering PreProcessing Package
}
\author{
Stefan Dentro, \email{sd11@sanger.ac.uk}
}
\keyword{package}

