% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{dpIn2vcf}
\alias{dpIn2vcf}
\title{DPClust input file to vcf}
\usage{
dpIn2vcf(vcf_infile, dpIn_file, vcf_outfile, fai_file, ign_file,
  genome = "hg19")
}
\arguments{
\item{vcf_infile}{Filename of the VCF file to use as a base}

\item{dpIn_file}{Filename of a DP input file}

\item{vcf_outfile}{Filename of the output file}

\item{fai_file}{Path to a reference genome index containing chromosome names}

\item{ign_file}{Path to a file containing contigs to ignore}

\item{genome}{Specify the reference genome for reading in the VCF}
}
\description{
Transform a dirichlet input file into a VCF with the same info. It filters out mutations in areas that are not contained in the supplied genome index (fai file) or are contained in the ignore file (ign file)
It takes the DP input file created by runGetDirichletProcessInfo and combines the columns with the vcf file supplied. Finally it gzips and indexes the file
}
\author{
sd11
}
