% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelecount.R
\name{dumpCounts.Broad}
\alias{dumpCounts.Broad}
\title{Dump allele counts from vcf - Broad ICGC pancancer pipeline}
\usage{
dumpCounts.Broad(
  vcf_infile,
  tumour_outfile,
  normal_outfile = NA,
  refence_genome = "hg19",
  samplename = NA
)
}
\arguments{
\item{vcf_infile}{The vcf file to read in}

\item{tumour_outfile}{File to save the tumour counts to}

\item{normal_outfile}{Optional parameter specifying where the normal output should go}

\item{refence_genome}{Optional parameter specifying the reference genome build used}

\item{samplename}{Optional parameter specifying the samplename to be used for matching the right column in the VCF}
}
\description{
Dump allele counts stored in the sample columns of the VCF file. Output will go into a file
supplied as tumour_outfile and optionally normal_outfile. It will be a fully formatted
allele counts file as returned by alleleCounter.
}
\author{
sd11
}
