% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{dumpCounts.DKFZ}
\alias{dumpCounts.DKFZ}
\title{Dump allele counts from vcf - DKFZ pipeline}
\usage{
dumpCounts.DKFZ(vcf_infile, tumour_outfile, normal_outfile = NA,
  refence_genome = "hg19")
}
\arguments{
\item{vcf_infile}{The vcf file to read in}

\item{tumour_outfile}{File to save the tumour counts to}

\item{normal_outfile}{Optional parameter specifying where the normal output should go}

\item{refence_genome}{Optional parameter specifying the reference genome build used}
}
\description{
Dump allele counts stored in the sample columns of the VCF file. Output will go into a file
supplied as tumour_outfile and optionally normal_outfile. It will be a fully formatted
allele counts file as returned by alleleCounter.
}
\author{
sd11
}

