% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{dumpCounts.ICGCconsensus}
\alias{dumpCounts.ICGCconsensus}
\title{Dump allele counts from vcf - ICGC pancancer consensus pipeline}
\usage{
dumpCounts.ICGCconsensus(vcf_infile, tumour_outfile, normal_outfile = NA,
  refence_genome = "hg19", samplename = NA)
}
\arguments{
\item{vcf_infile}{The vcf file to read in}

\item{tumour_outfile}{File to save the tumour counts to}

\item{normal_outfile}{Optional parameter specifying where the normal output should go}

\item{refence_genome}{Optional parameter specifying the reference genome build used}

\item{samplename}{Optional parameter specifying the samplename to be used for matching the right column in the VCF}
}
\description{
Dump allele counts stored in the info column of the VCF file. Output will go into a file
supplied as tumour_outfile. It will be a fully formatted allele counts file as returned
by alleleCounter. There are no counts for the matched normal.
}
\author{
sd11
}

