% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelecount.R
\name{dumpCounts.mutect}
\alias{dumpCounts.mutect}
\title{Dump allele counts from vcf - Mutect}
\usage{
dumpCounts.mutect(
  vcf_infile,
  tumour_outfile,
  samplename,
  normal_outfile = NA,
  refence_genome = "hg19"
)
}
\arguments{
\item{vcf_infile}{The vcf file to read in}

\item{tumour_outfile}{File to save the tumour counts to}

\item{samplename}{Parameter specifying the samplename to be used for matching the right column in the VCF}

\item{normal_outfile}{Optional parameter specifying where the normal output should go}

\item{refence_genome}{Optional parameter specifying the reference genome build used}
}
\description{
Dump allele counts stored in the info column of the VCF file. Output will go into a file
supplied as tumour_outfile. It will be a fully formatted allele counts file as returned 
by alleleCounter.
}
\author{
sd11
}
