% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumber.R
\name{extend_bb_cn_categories}
\alias{extend_bb_cn_categories}
\title{This function extends existing categories by a list of sizes}
\usage{
extend_bb_cn_categories(
  infile,
  outfile,
  size_categories = c(`3` = 10^3, `4` = 10^4, `5` = 10^5, `6` = 10^6, `7` = 10^7, `8` =
    10^8, `9` = 10^9, `10` = 10^10)
)
}
\arguments{
\item{infile}{string that points to a collated Battenberg subclones file}

\item{outfile}{string where the output will be written.}

\item{size_categories}{named vector with size options to consider for each CNA category}
}
\description{
This function extends existing categories by a list of sizes
}
\author{
sd11
}
