% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{filterForDeaminase}
\alias{filterForDeaminase}
\title{Convenience function that filters a with tri-nucleotide context annotated list of loci for
cytosine deaminase signature, or C>T at CpG.}
\usage{
filterForDeaminase(loci_file, outfile, trinucleotide_column = 5,
  ref_genome = "hg19")
}
\arguments{
\item{outfile}{Filepath to where to store the output.}

\item{trinucleotide_column}{Integer representing the column within the input file that contains the context}

\item{signature_anno_loci_file}{Filepath to a with tri-nucleotide context annotated loci}
}
\description{
Convenience function that filters a with tri-nucleotide context annotated list of loci for
cytosine deaminase signature, or C>T at CpG.
}
\author{
sd11
}

