% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{filterForSignature}
\alias{filterForSignature}
\title{Filter a with tri-nucleotide context annotated list of loci for a particular signature
supplied as a regex like so: (CAG)|(CTG)|(GAC)|(GTC).}
\usage{
filterForSignature(
  signature_anno_loci_file,
  signature_regex,
  outfile,
  trinucleotide_column = 5,
  alt_alleles = NA,
  alt_allele_column = 4
)
}
\arguments{
\item{signature_anno_loci_file}{Filepath to a with tri-nucleotide context annotated loci}

\item{signature_regex}{A regex that captures the signature to keep.}

\item{outfile}{Filepath to where to store the output.}

\item{trinucleotide_column}{Integer representing the column within the input file that contains the context}

\item{alt_alleles}{An optional vector with reference alleles to allow.}

\item{alt_allele_column}{The column in the annotated loci file that contains the reference base.}
}
\description{
Filter a with tri-nucleotide context annotated list of loci for a particular signature
supplied as a regex like so: (CAG)|(CTG)|(GAC)|(GTC).
}
\author{
sd11
}
