% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumber.R
\name{get_bb_cn_catalog}
\alias{get_bb_cn_catalog}
\title{Count how often each of the categories is present in this sample}
\usage{
get_bb_cn_catalog(
  infile,
  outfile,
  size_categories = c(`3` = 10^3, `4` = 10^4, `5` = 10^5, `6` = 10^6, `7` = 10^7, `8` =
    10^8, `9` = 10^9, `10` = 10^10)
)
}
\arguments{
\item{infile}{string pointing to a collated Battenberg subclones file with extended CNA categories}

\item{outfile}{string pointing to where the catalog should be written}

\item{size_categories}{named vector with the size options to consider. This should be the same vector given to extend_bb_cn_categories}
}
\description{
Count how often each of the categories is present in this sample
}
\author{
sd11
}
