% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelecount.R
\name{getAlleleCounts.ICGC_consensus_indel}
\alias{getAlleleCounts.ICGC_consensus_indel}
\title{Dump allele counts in ICGC consensus indel format}
\usage{
getAlleleCounts.ICGC_consensus_indel(v, dummy_alt_allele = "A",
  dummy_ref_allele = "C")
}
\arguments{
\item{v}{The vcf file}

\item{dummy_alt_allele}{Dummy base to use to encode the alt allele, the counts will be saved in the corresponding column}

\item{dummy_ref_allele}{Dummy base to use to encode the ref allele, the counts will be saved in the corresponding column}
}
\value{
An array with 4 columns: Counts for A, C, G, T
}
\description{
This function fetches allele counts from the info field in the VCF file.
}
\author{
sd11
}
