% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelecount.R
\name{getAlleleCounts.ICGC_consensus_snv}
\alias{getAlleleCounts.ICGC_consensus_snv}
\title{Dump allele counts in ICGC consensus SNV format}
\usage{
getAlleleCounts.ICGC_consensus_snv(v)
}
\arguments{
\item{v}{The vcf file}
}
\value{
An array with 4 columns: Counts for A, C, G, T
}
\description{
This function fetches allele counts from the info field in the VCF file.
}
\author{
sd11
}
