% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelecount.R
\name{getAlleleCounts.mutect}
\alias{getAlleleCounts.mutect}
\title{Dump allele counts in Mutect format}
\usage{
getAlleleCounts.mutect(v, sample_col)
}
\arguments{
\item{v}{The vcf file}

\item{sample_col}{The column in which the counts are. If it's the first sample mentioned in the vcf this would be sample_col 1}
}
\value{
An array with 4 columns: Counts for A, C, G, T
}
\description{
This function fetches allele counts from the info field in the VCF file.
}
\author{
sd11
}
