% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{getAlleleCounts.Sanger}
\alias{getAlleleCounts.Sanger}
\title{Helper function that dumps the allele counts from a Sanger pipeline VCF file}
\usage{
getAlleleCounts.Sanger(v, sample_col)
}
\arguments{
\item{v}{The vcf file}

\item{sample_col}{The column in which the counts are. If it's the first sample mentioned in the vcf this would be sample_col 1}
}
\description{
Helper function that dumps the allele counts from a Sanger pipeline VCF file
}
\author{
sd11
}

