% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelecount.R
\name{getAltAllele}
\alias{getAltAllele}
\title{Obtain alt allele for all variants as a vector}
\usage{
getAltAllele(vcf)
}
\arguments{
\item{vcf}{A VCF object}
}
\value{
A vector with the alt allele
}
\description{
Dumps the alt allele in a vector. It takes only the first element if multiple alt alleles are listed
}
\author{
sd11
}
