% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{getCountsNormal}
\alias{getCountsNormal}
\title{Returns an allele counts table for the normal sample}
\usage{
getCountsNormal(v, centre = "sanger")
}
\arguments{
\item{v}{The vcf file}

\item{centre}{The sequencing centre of which pipeline the vcf file originates}
}
\description{
Returns an allele counts table for the normal sample
}
\author{
sd11
}

