% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{getCountsTumour}
\alias{getCountsTumour}
\title{Dump allele counts from vcf for tumour}
\usage{
getCountsTumour(v, centre = "sanger")
}
\arguments{
\item{v}{The vcf file}

\item{centre}{The sequencing centre of which pipeline the vcf file originates}
}
\description{
Returns an allele counts table for the tumour sample
}
\author{
sd11
}

