% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelecount.R
\name{getRefAllele}
\alias{getRefAllele}
\title{Obtain ref allele for all variants as a vector}
\usage{
getRefAllele(vcf)
}
\arguments{
\item{vcf}{A VCF object}
}
\value{
A vector with the ref allele
}
\description{
Dumps the ref allele in a vector. It takes only the first element if multiple alt alleles are listed
}
\author{
davidwedge
}
