% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{getTrinucleotideContext}
\alias{getTrinucleotideContext}
\title{Obtain tri-nucleotide context. It reads in the supplied loci file and querries
the provided reference genome fasta for the context. Finally it writes out the
loci with annotated context.}
\usage{
getTrinucleotideContext(loci_file, outfile, ref_genome)
}
\arguments{
\item{loci_file}{A 4 column loci file with chrom, position, reference allele and alt allele}

\item{outfile}{Where to write the output}

\item{ref_genome}{Full path to a reference genome Fasta file}
}
\description{
Obtain tri-nucleotide context. It reads in the supplied loci file and querries
the provided reference genome fasta for the context. Finally it writes out the
loci with annotated context.
}
\author{
sd11
}
