% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kataegis.R
\name{identifyKataegis}
\alias{identifyKataegis}
\title{Identify Kataegis events in SNV mutation data}
\usage{
identifyKataegis(
  samplename,
  dpInfile,
  outdir = ".",
  gamma_param = 25,
  kmin = 2,
  kataegis.threshold = NA,
  minMuts = 6,
  logScale = F,
  makePlots = F,
  removeFromDPin = F
)
}
\arguments{
\item{samplename}{Samplename used when writing output files}

\item{dpInfile}{A DP input file}

\item{outdir}{Directory where output will be written}

\item{gamma_param}{Gamma parameter to be used for segmentation}

\item{kmin}{Kmin parameter to be used for segmentation}

\item{kataegis.threshold}{Intermutation distance, if NA will be set to: (a) 900000 > SNVs: 100, (b) 500000 > SNVs: 250, (c) 100000 > SNVs, (d) otherwise 1000 (Default NA)}

\item{minMuts}{Minimum number of mutations within kataegis.threshold to call Kataegis}

\item{logScale}{Transform intermutation distance to logscale}

\item{makePlots}{Make a figure for each Kataegis event}

\item{removeFromDPin}{Remove SNVs identified as part of a Kataegis event from the DP input file}
}
\description{
Identify Kataegis events in SNV mutation data
}
\author{
dw9
}
