% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LoadData.R
\name{load.data}
\alias{load.data}
\title{THIS FUNCTION IS DEPRECATED - SHOULD BE MERGED WITH THE ONE IN DPCLUST}
\usage{
load.data(datpath, samplename, list_of_data_files, cellularity, Chromosome,
  position, WT.count, mut.count, subclonal.CN, no.chrs.bearing.mut,
  mutation.copy.number, subclonal.fraction, data_file_suffix)
}
\arguments{
\item{datpath}{Path to where the input datafiles can be found, type string.}

\item{samplename}{A unique identifier for this particular individual/experiment, type string.}

\item{list_of_data_files}{A list of file names that are available in datpath, list of strings.}

\item{cellularity}{Vector containing the cellularity estimate for each sample in this set, vector of floats.}

\item{Chromosome}{The name of the column containing the chromosome information in the input, type string.}

\item{position}{The name of the column containing the position information in the input, type string.}

\item{WT.count}{The name of the column containing the wild type allele count information in the input, type string.}

\item{mut.count}{The name of the column containing the mutation allele count information in the input, type string.}

\item{subclonal.CN}{The name of the column containing the total subclonal copynumber information in the input, type string.}

\item{no.chrs.bearing.mut}{The name of the column containing the number of chromsomes bearing the mutation information in the input, type string.}

\item{mutation.copy.number}{Mutation copy number}

\item{subclonal.fraction}{Subclonal fraction}

\item{data_file_suffix}{A suffix to be appended to each item in list_of_data_files to complete the filenames (for example dirichletInput.R), type string.}
}
\value{
A named list of matrices and vectors with the following items: chromosome, position, WTCount, mutCount, totalCopyNumber, copyNumberAdjustment, non.deleted.muts, kappa, mutation.copy.number, subclonal.fraction, removed_indices, chromosome.not.filtered, mut.position.not.filtered
}
\description{
load.data
}
\details{
Load data files for a specific set of samples belonging to the same individual/experiment.
}
\author{
sd11
}

