% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{mut_cn_phasing}
\alias{mut_cn_phasing}
\title{Phase mutation to SNP/copy number}
\usage{
mut_cn_phasing(
  loci_file,
  phased_file,
  hap_file,
  bam_file,
  bai_file,
  outfile,
  max_distance
)
}
\arguments{
\item{loci_file}{A list of loci}

\item{phased_file}{The .BAFsegmented.txt output file from Battenberg}

\item{hap_file}{Path to the _allHaplotypeInfo.txt Battenberg output file to be used}

\item{bam_file}{Full path to the bam file}

\item{bai_file}{Full path to the bai file}

\item{outfile}{File to save the output}

\item{max_distance}{The max distance of a mutation and SNP can be apart to be considered for phasing}
}
\description{
Run mutation to copy number phasing. This function requires the Linkage_pull.pl script in $PATH.
Note: This function should either be run separately per chromosome and then combined with \code{\link{concat_files}}
or on all chromsomes in one go, but then the _allHaplotypeInfo.txt Battenberg files need to be concatenated first.
}
\author{
sd11, dw9
}
