% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{mut_mut_phasing}
\alias{mut_mut_phasing}
\title{Phase mutation to mutation}
\usage{
mut_mut_phasing(loci_file, phased_file, bam_file, bai_file, max_distance)
}
\arguments{
\item{loci_file}{A list of loci}

\item{phased_file}{File to save the output}

\item{bam_file}{Full path to the bam file}

\item{bai_file}{Full path to the bai file}

\item{max_distance}{The max distance of a mutation and SNP can be apart to be considered for phasing}
}
\description{
Run mutation to mutation phasing. This function requires the Linkage_pull.pl script in $PATH.
}
\author{
sd11, dw9
}
