% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interconvertMutationBurdens.R
\name{mutationBurdenToMutationCopyNumber}
\alias{mutationBurdenToMutationCopyNumber}
\title{Mutation burden to mutation copy number}
\usage{
mutationBurdenToMutationCopyNumber(
  burden,
  totalCopyNumber,
  cellularity,
  normalCopyNumber = rep(2, length(burden))
)
}
\arguments{
\item{burden}{A vector containing mutation burdens}

\item{totalCopyNumber}{A vector with total tumour copynumber}

\item{cellularity}{Float with the cellularity of this sample}

\item{normalCopyNumber}{A vector with the total copy number of the normal}
}
\value{
A vector with mutation copy number
}
\description{
Function to convert mutation burdens into mutation copy number
}
\author{
dw9
}
