% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interconvertMutationBurdens.R
\name{mutationCopyNumberToMutationBurden}
\alias{mutationCopyNumberToMutationBurden}
\title{Mutation copy number to mutation burden}
\usage{
mutationCopyNumberToMutationBurden(
  copyNumber,
  totalCopyNumber,
  cellularity,
  normalCopyNumber = rep(2, length(copyNumber))
)
}
\arguments{
\item{copyNumber}{A vector containing mutation copy number}

\item{totalCopyNumber}{A vector with total tumour copynumber}

\item{cellularity}{Float with the cellularity of this sample}

\item{normalCopyNumber}{A vector with the total copy number of the normal}
}
\value{
A vector with mutation burdens
}
\description{
Function to convert mutation copy number to mutation burden
}
\author{
dw9
}
