% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitycontrol.R
\name{plot_ccf_hist}
\alias{plot_ccf_hist}
\title{Plot cancer cell fraction histogram}
\usage{
plot_ccf_hist(data, samplename, outdir)
}
\arguments{
\item{data}{A DPClust input table}

\item{samplename}{Name of the sample}

\item{outdir}{Directory where the figure is to be stored}
}
\description{
Plot cancer cell fraction histogram
}
\author{
sd11
}
