% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitycontrol.R
\name{plot_mcn_hist}
\alias{plot_mcn_hist}
\title{Plot mutation copy number histogram}
\usage{
plot_mcn_hist(data, samplename, outdir)
}
\arguments{
\item{data}{A DPClust input table}

\item{samplename}{Name of the sample}

\item{outdir}{Directory where the figure is to be stored}
}
\description{
Plot mutation copy number histogram
}
\author{
sd11
}
