% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitycontrol.R
\name{plot_vaf_hist}
\alias{plot_vaf_hist}
\title{Plot allele frequency histogram}
\usage{
plot_vaf_hist(data, samplename, outdir)
}
\arguments{
\item{data}{A DPClust input table}

\item{samplename}{Name of the sample}

\item{outdir}{Directory where the figure is to be stored}
}
\description{
Plot allele frequency histogram
}
\author{
sd11
}
