% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{split_by_chrom}
\alias{split_by_chrom}
\title{Split a file per chromosome}
\usage{
split_by_chrom(infile, prefix, postfix, outdir, chrom_file)
}
\arguments{
\item{infile}{The file to be split}

\item{prefix}{Prefix of the output file}

\item{postfix}{Postfix of the output file}

\item{outdir}{Directory where the output files are to be written}

\item{chrom_file}{A simple list of chromosomes to be considered}
}
\description{
Convenience function to split an input file per chromosome. All it requires is that
the infile has as first column chromosome specification. The output files will be named
outdir/prefixCHROMNUMBERpostfix
}
\author{
sd11
}

