% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumber.R
\name{collate_bb_subclones}
\alias{collate_bb_subclones}
\title{This function takes a Battenberg styled data.frame, pulls out the A copy number profile,
then it calculates the total CN and a weighted ploidy before it subsets the input data
by columns: "chr", "startpos", "endpos", "nMaj1_A", "nMin1_A", "frac1_A", 
"nMaj2_A", "nMin2_A", "frac2_A", "SDfrac_A".}
\usage{
collate_bb_subclones(
  samplename,
  cndata,
  gender,
  outfile,
  ploidy = NA,
  exclude_sex_chroms = F
)
}
\arguments{
\item{samplename}{A String containing the samplename}

\item{cndata}{A data.frame in Battenberg subclones format}

\item{gender}{Specify either the string male or female}

\item{outfile}{A String with the full path to where the output should be written}

\item{ploidy}{An optional parameter that is used to call gains/losses against. If higher than the ploidy a segment is considered a gain, lower a loss. When left as NA the sample ploidy is used (Default NA).}

\item{exclude_sex_chroms}{Optional parameter whether to exclude sex chromosomes (Default FALSE)}
}
\description{
Then logic is applied for classifying copy number aberrations into
a series of classes. There are two main groups: Clonal (starting with c) and
Subclonal (starting with s).
}
\details{
A CNA can be:
* HD : Homozygous deletion
* LOH : Loss of heterozygosity
* Amp : Amplification
* Gain : Gain
* NoCNV : Normal copy number without aberrations
* Loss : Loss

The copy number segments, classification, weighted ploidy and samplename are saved
to the specified output file
}
\author{
tjm
}
