% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{filterForDeaminase}
\alias{filterForDeaminase}
\title{Convenience function that filters a with tri-nucleotide context annotated list of loci for
cytosine deaminase signature.}
\usage{
filterForDeaminase(signature_anno_loci_file, outfile,
  trinucleotide_column = 5)
}
\arguments{
\item{signature_anno_loci_file}{Filepath to a with tri-nucleotide context annotated loci}

\item{outfile}{Filepath to where to store the output.}

\item{trinucleotide_column}{Integer representing the column within the input file that contains the context}
}
\description{
Convenience function that filters a with tri-nucleotide context annotated list of loci for
cytosine deaminase signature.
}
\author{
sd11
}

