% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{filterForSignature}
\alias{filterForSignature}
\title{#Checks each tri-nucleotide context whether its CAG or CTG
#}
\usage{
filterForSignature(signature_anno_loci_file, signature_regex, outfile,
  trinucleotide_column = 5)
}
\arguments{
\item{signature_anno_loci_file}{Filepath to a with tri-nucleotide context annotated loci}

\item{signature_regex}{A regex that captures the signature to keep.}

\item{outfile}{Filepath to where to store the output.}

\item{trinucleotide_column}{Integer representing the column within the input file that contains the context}

\item{vector_of_trinucleotides}{A vector containing the base to check and its immediate neighbors
#}
}
\description{
#Checks each tri-nucleotide context whether its CAG or CTG
#
}
\author{
sd11
Filter a with tri-nucleotide context annotated list of loci for a particular signature
supplied as a regex like so: (CAG)|(CTG)|(GAC)|(GTC).

sd11
}

