% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{parseIgnore}
\alias{parseIgnore}
\title{Convenience function that parses an ignore file. This file
is expected to have a single column with just chromosome names}
\usage{
parseIgnore(ignore_file)
}
\arguments{
\item{ignore_file}{The file specifying to be ignored chromosomes}
}
\value{
A data frame with a single column named "chromosome"
}
\description{
Convenience function that parses an ignore file. This file
is expected to have a single column with just chromosome names
}
\author{
sd11
}

