% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{runGetDirichletProcessInfo}
\alias{runGetDirichletProcessInfo}
\title{Create the DPClust input file}
\usage{
runGetDirichletProcessInfo(
  loci_file,
  allele_frequencies_file,
  cellularity_file,
  subclone_file,
  gender,
  SNP.phase.file,
  mut.phase.file,
  output_file
)
}
\arguments{
\item{loci_file}{Simple four column file with chromosome, position, reference allele and alternative allele}

\item{allele_frequencies_file}{Output file from alleleCounter on the specified loci}

\item{cellularity_file}{Full path to a Battenberg rho_and_psi output file}

\item{subclone_file}{Full path to a Battenberg subclones.txt output file}

\item{gender}{Specify male or female}

\item{SNP.phase.file}{Output file from mut_mut_phasing, supply NA (as char) when not available}

\item{mut.phase.file}{Output file from mut_cn_phasing, supply NA (as char) when not available}

\item{output_file}{Name of the output file}
}
\description{
Function that takes allele counts and a copy number profile to estimate mutation copy number,
cancer cell fraction and multiplicity for each point mutation.
}
\author{
sd11
}
